refactor: share components between in-app and home-screen widgets
