+++ Bitkit/Components/Widgets/BlocksWidget.swift
@@ -46,74 +46,6 @@ struct BlocksWidget: View {
     }
 }
 
-// MARK: - Wide layout (in-app + 343-wide carousel page + .systemMedium / .systemLarge OS widget)
-
-struct BlocksWidgetWideContent: View {
-    let data: CachedBlock
-    let options: BlocksWidgetOptions
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 12) {
-            ForEach(options.enabledFields, id: \.self) { field in
-                BlocksWidgetWideRow(field: field, value: field.value(from: data))
-            }
-        }
-        .frame(maxWidth: .infinity, alignment: .leading)
-    }
-}
-
-private struct BlocksWidgetWideRow: View {
-    let field: BlocksWidgetField
-    let value: String
-
-    var body: some View {
-        HStack(alignment: .center, spacing: 8) {
-            Image(field.iconName)
-                .resizable()
-                .renderingMode(.template)
-                .foregroundColor(.brandAccent)
-                .frame(width: 20, height: 20)
-
-            BodyMText(field.label, textColor: .white80)
-                .lineLimit(1)
-                .frame(maxWidth: .infinity, alignment: .leading)
-
-            BodyMSBText(value)
-                .lineLimit(1)
-                .truncationMode(.middle)
-        }
-    }
-}
-
-// MARK: - Compact layout (small carousel preview + 163×192 OS small widget)
-
-struct BlocksWidgetCompactContent: View {
-    let data: CachedBlock
-    let options: BlocksWidgetOptions
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 16) {
-            ForEach(options.enabledFields, id: \.self) { field in
-                HStack(alignment: .center, spacing: 8) {
-                    Image(field.iconName)
-                        .resizable()
-                        .renderingMode(.template)
-                        .foregroundColor(.brandAccent)
-                        .frame(width: 20, height: 20)
-
-                    BodySSBText(field.value(from: data))
-                        .lineLimit(1)
-                        .truncationMode(.middle)
-                }
-            }
-        }
-        .padding(16)
-        .frame(maxWidth: .infinity, maxHeight: .infinity, alignment: .topLeading)
-        .background(Color.gray6)
-        .cornerRadius(16)
-    }
-}
-
 #Preview {
     BlocksWidget()
         .padding()
