+++ Bitkit/Components/Widgets/PriceWidget.swift
@@ -1,4 +1,3 @@
-import Charts
 import SwiftUI
 
 /// Displays Bitcoin price for the user's selected trading pair and timeframe.
@@ -52,135 +51,6 @@ struct PriceWidget: View {
 
     private func fetchPriceData() {
         viewModel.fetchPriceData(pairs: [options.selectedPair], period: options.selectedPeriod)
-    }
-}
-
-// MARK: - Wide layout (in-app + carousel page)
-
-struct PriceWidgetWideContent: View {
-    let data: PriceData
-    let period: GraphPeriod
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 8) {
-            VStack(alignment: .leading, spacing: 4) {
-                HStack(alignment: .center, spacing: 16) {
-                    CaptionMText("\(data.name)  \(period.rawValue)", textColor: .textSecondary)
-                        .textCase(.uppercase)
-                        .frame(maxWidth: .infinity, alignment: .leading)
-
-                    TitleText(
-                        data.change.formatted,
-                        textColor: data.change.isPositive ? .greenAccent : .redAccent
-                    )
-                    .lineLimit(1)
-                    .accessibilityIdentifier("price_card_pair_change_\(data.name)")
-                }
-                .accessibilityIdentifier("PriceWidgetRow-\(data.name)")
-
-                Text(data.price)
-                    .font(Fonts.bold(size: 34))
-                    .foregroundColor(.textPrimary)
-                    .lineLimit(1)
-                    .minimumScaleFactor(0.7)
-                    .frame(maxWidth: .infinity, alignment: .leading)
-                    .accessibilityIdentifier("price_card_pair_price_\(data.name)")
-            }
-
-            PriceChart(values: data.pastValues, isPositive: data.change.isPositive)
-                .frame(height: 48)
-                .accessibilityIdentifier("price_card_chart")
-        }
-        .frame(maxWidth: .infinity, alignment: .leading)
-    }
-}
-
-// MARK: - Compact layout (small carousel preview only)
-
-struct PriceWidgetCompactContent: View {
-    let data: PriceData
-    let period: GraphPeriod
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 16) {
-            VStack(alignment: .leading, spacing: 8) {
-                HStack(spacing: 0) {
-                    CaptionMText(data.name, textColor: .textSecondary)
-                        .textCase(.uppercase)
-                    Spacer(minLength: 0)
-                    CaptionMText(period.rawValue, textColor: .textSecondary)
-                        .textCase(.uppercase)
-                }
-                .accessibilityIdentifier("price_card_small_pair_row_\(data.name)")
-
-                Text(data.price)
-                    .font(Fonts.bold(size: 22))
-                    .foregroundColor(.textPrimary)
-                    .lineLimit(1)
-                    .minimumScaleFactor(0.7)
-                    .accessibilityIdentifier("price_card_small_pair_price_\(data.name)")
-
-                BodySSBText(
-                    data.change.formatted,
-                    textColor: data.change.isPositive ? .greenAccent : .redAccent
-                )
-                .lineLimit(1)
-                .accessibilityIdentifier("price_card_small_pair_change_\(data.name)")
-            }
-
-            PriceChart(values: data.pastValues, isPositive: data.change.isPositive)
-                .frame(height: 64)
-                .accessibilityIdentifier("price_card_small_chart")
-        }
-        .padding(16)
-        .frame(maxWidth: .infinity, maxHeight: .infinity, alignment: .topLeading)
-        .background(Color.gray6)
-        .cornerRadius(16)
-    }
-}
-
-// MARK: - Chart
-
-struct PriceChart: View {
-    let values: [Double]
-    let isPositive: Bool
-
-    private let lineWidth: CGFloat = 1.3
-
-    private var normalizedValues: [Double] {
-        guard values.count > 1 else { return values }
-
-        let minValue = values.min() ?? 0
-        let maxValue = values.max() ?? 0
-        let range = maxValue - minValue
-
-        guard range > 0 else { return values.map { _ in 0.5 } }
-
-        return values.map { value in
-            let normalized = (value - minValue) / range
-            return 0.15 + (normalized * 0.7)
-        }
-    }
-
-    private var lineColor: Color {
-        isPositive ? .greenAccent : .redAccent
-    }
-
-    var body: some View {
-        Chart {
-            ForEach(Array(normalizedValues.enumerated()), id: \.offset) { index, value in
-                LineMark(
-                    x: .value("Index", index),
-                    y: .value("Price", value)
-                )
-                .foregroundStyle(lineColor)
-                .lineStyle(StrokeStyle(lineWidth: lineWidth))
-                .interpolationMethod(.catmullRom)
-            }
-        }
-        .chartXAxis(.hidden)
-        .chartYAxis(.hidden)
-        .chartYScale(domain: 0.1 ... 0.9)
     }
 }
