+++ BitkitWidget/BlocksHomeScreenWidget.swift
@@ -93,9 +93,13 @@ struct BlocksHomeScreenWidgetEntryView: View {
 
     var entry: BlocksWidgetProvider.Entry
 
+    private var palette: WidgetPalette {
+        WidgetPalette(renderingMode: widgetRenderingMode)
+    }
+
     var body: some View {
         content
-            .containerBackground(for: .widget) { backgroundView }
+            .containerBackground(for: .widget) { palette.background }
     }
 
     @ViewBuilder
@@ -105,9 +109,10 @@ struct BlocksHomeScreenWidgetEntryView: View {
         } else if let block = entry.block {
             switch widgetFamily {
             case .systemSmall:
-                compactLayout(block: block)
+                BlocksWidgetCompactContent(data: block, options: entry.options)
             default:
-                wideLayout(block: block, fields: entry.options.enabledFields)
+                BlocksWidgetWideContent(data: block, options: entry.options)
+                    .frame(maxWidth: .infinity, maxHeight: .infinity, alignment: .topLeading)
             }
         } else {
             ProgressView()
@@ -115,103 +120,9 @@ struct BlocksHomeScreenWidgetEntryView: View {
         }
     }
 
-    // MARK: - Layouts
-
-    /// Candidate row spacings, from roomiest to tightest. `ViewThatFits` picks the first that
-    /// fits the available height, so larger screens keep 16pt while iPhone SE falls back to less.
-    private static let rowSpacings: [CGFloat] = [16, 10, 6, 2]
-
-    /// Compact (`.systemSmall`): icon + value rows for the selected fields.
-    private func compactLayout(block: CachedBlock) -> some View {
-        ViewThatFits(in: .vertical) {
-            ForEach(Self.rowSpacings, id: \.self) { spacing in
-                compactStack(block: block, spacing: spacing)
-            }
-        }
-        .frame(maxWidth: .infinity, maxHeight: .infinity, alignment: .topLeading)
-    }
-
-    private func compactStack(block: CachedBlock, spacing: CGFloat) -> some View {
-        VStack(alignment: .leading, spacing: spacing) {
-            ForEach(entry.options.enabledFields, id: \.self) { field in
-                HStack(alignment: .center, spacing: 8) {
-                    iconImage(field: field)
-                    Text(field.value(from: block))
-                        .font(Fonts.semiBold(size: 15))
-                        .foregroundColor(titleTextColor)
-                        .lineLimit(1)
-                        .truncationMode(.middle)
-                        .widgetAccentable()
-                }
-            }
-        }
-        .frame(maxWidth: .infinity, alignment: .topLeading)
-    }
-
-    /// Wide layout (`.systemMedium`): icon + label + value rows for the selected fields.
-    private func wideLayout(block: CachedBlock, fields: [BlocksWidgetField]) -> some View {
-        ViewThatFits(in: .vertical) {
-            ForEach(Self.rowSpacings, id: \.self) { spacing in
-                wideStack(block: block, fields: fields, spacing: spacing)
-            }
-        }
-        .frame(maxWidth: .infinity, maxHeight: .infinity, alignment: .topLeading)
-    }
-
-    private func wideStack(block: CachedBlock, fields: [BlocksWidgetField], spacing: CGFloat) -> some View {
-        VStack(alignment: .leading, spacing: spacing) {
-            ForEach(fields, id: \.self) { field in
-                HStack(alignment: .center, spacing: 8) {
-                    iconImage(field: field)
-                    Text(field.label)
-                        .font(Fonts.regular(size: 17))
-                        .foregroundColor(labelTextColor)
-                        .lineLimit(1)
-                        .frame(maxWidth: .infinity, alignment: .leading)
-                    Text(field.value(from: block))
-                        .font(Fonts.semiBold(size: 17))
-                        .foregroundColor(titleTextColor)
-                        .lineLimit(1)
-                        .truncationMode(.middle)
-                        .widgetAccentable()
-                }
-            }
-        }
-        .frame(maxWidth: .infinity, alignment: .topLeading)
-    }
-
-    private func iconImage(field: BlocksWidgetField) -> some View {
-        Image(field.iconName)
-            .resizable()
-            .renderingMode(.template)
-            .foregroundColor(iconColor)
-            .frame(width: 20, height: 20)
-            .widgetAccentable()
-    }
-
     private var errorView: some View {
-        Text("Couldn’t load blocks data.")
-            .font(Fonts.regular(size: 13))
-            .foregroundColor(labelTextColor)
+        BodySText(t("widgets__blocks__error"), textColor: palette.secondary)
             .frame(maxWidth: .infinity, maxHeight: .infinity, alignment: .topLeading)
-    }
-
-    // MARK: - Colors
-
-    private var backgroundView: some View {
-        widgetRenderingMode == .fullColor ? Color.gray6 : Color.clear
-    }
-
-    private var titleTextColor: Color {
-        widgetRenderingMode == .fullColor ? .white : .primary
-    }
-
-    private var labelTextColor: Color {
-        widgetRenderingMode == .fullColor ? .white.opacity(0.8) : .secondary
-    }
-
-    private var iconColor: Color {
-        widgetRenderingMode == .fullColor ? .brandAccent : .primary
     }
 }
