+++ BitkitWidget/NewsHomeScreenWidget.swift
@@ -133,6 +133,10 @@ struct NewsHomeScreenWidgetEntryView: View {
 
     var entry: NewsWidgetProvider.Entry
 
+    private var palette: WidgetPalette {
+        WidgetPalette(renderingMode: widgetRenderingMode)
+    }
+
     var body: some View {
         Group {
             if let url = articleURL {
@@ -142,7 +146,7 @@ struct NewsHomeScreenWidgetEntryView: View {
             }
         }
         .widgetURL(articleURL)
-        .containerBackground(for: .widget) { backgroundView }
+        .containerBackground(for: .widget) { palette.background }
     }
 
     private var articleURL: URL? {
@@ -157,9 +161,19 @@ struct NewsHomeScreenWidgetEntryView: View {
         } else if let article = entry.article {
             switch widgetFamily {
             case .systemSmall:
-                compactLayout(article: article)
+                NewsWidgetCompactContent(
+                    title: article.title,
+                    timeAgo: entry.timeAgo,
+                    options: entry.options
+                )
             default:
-                wideLayout(article: article)
+                NewsWidgetWideContent(
+                    title: article.title,
+                    publisher: article.publisher,
+                    timeAgo: entry.timeAgo,
+                    options: entry.options
+                )
+                .frame(maxWidth: .infinity, maxHeight: .infinity, alignment: .topLeading)
             }
         } else {
             ProgressView()
@@ -167,96 +181,9 @@ struct NewsHomeScreenWidgetEntryView: View {
         }
     }
 
-    // MARK: - Compact (small widget — 163×192)
-
-    private func compactLayout(article: CachedNewsArticle) -> some View {
-        VStack(alignment: .leading, spacing: 0) {
-            titleText(article.title)
-                .frame(maxWidth: .infinity, alignment: .leading)
-
-            Spacer(minLength: 8)
-
-            if entry.options.showDate {
-                HStack {
-                    Spacer(minLength: 0)
-                    Text(entry.timeAgo)
-                        .font(Fonts.semiBold(size: 13))
-                        .tracking(0.4)
-                        .foregroundColor(secondaryTextColor)
-                        .lineLimit(1)
-                }
-            }
-        }
-        .frame(maxWidth: .infinity, maxHeight: .infinity, alignment: .topLeading)
-    }
-
-    // MARK: - Wide (medium widget — 343×118)
-
-    private func wideLayout(article: CachedNewsArticle) -> some View {
-        VStack(alignment: .leading, spacing: 0) {
-            titleText(article.title)
-                .frame(maxWidth: .infinity, alignment: .leading)
-
-            Spacer(minLength: 8)
-
-            if entry.options.showSource || entry.options.showDate {
-                HStack(alignment: .center, spacing: 8) {
-                    if entry.options.showSource {
-                        Text(article.publisher)
-                            .font(Fonts.semiBold(size: 13))
-                            .tracking(0.4)
-                            .foregroundColor(sourceTextColor)
-                            .lineLimit(1)
-                    }
-                    Spacer(minLength: 0)
-                    if entry.options.showDate {
-                        Text(entry.timeAgo)
-                            .font(Fonts.semiBold(size: 13))
-                            .tracking(0.4)
-                            .foregroundColor(secondaryTextColor)
-                            .lineLimit(1)
-                    }
-                }
-            }
-        }
-        .frame(maxWidth: .infinity, maxHeight: .infinity, alignment: .topLeading)
-    }
-
-    // MARK: - Sub-views
-
-    private func titleText(_ value: String) -> some View {
-        Text(value)
-            .font(Fonts.bold(size: 22))
-            .foregroundColor(titleTextColor)
-            .lineLimit(4)
-            .minimumScaleFactor(0.85)
-            .widgetAccentable()
-    }
-
     private var errorView: some View {
-        Text("Couldn’t load headlines.")
-            .font(Fonts.regular(size: 13))
-            .foregroundColor(secondaryTextColor)
+        BodySText(t("widgets__news__error"), textColor: palette.secondary)
             .frame(maxWidth: .infinity, maxHeight: .infinity, alignment: .topLeading)
-    }
-
-    // MARK: - Colors
-
-    private var backgroundView: some View {
-        widgetRenderingMode == .fullColor ? Color.gray6 : Color.clear
-    }
-
-    private var titleTextColor: Color {
-        widgetRenderingMode == .fullColor ? .white : .primary
-    }
-
-    private var sourceTextColor: Color {
-        guard widgetRenderingMode == .fullColor else { return .primary }
-        return .brandAccent
-    }
-
-    private var secondaryTextColor: Color {
-        widgetRenderingMode == .fullColor ? .white.opacity(0.64) : .secondary
     }
 }
